+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "theme-shadow-vscode" extension will be documented in
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.3] - 2019-12-30
+
+- Automate theme with new colors
+- Change color as per your requirement in [colors](themes/colors.mjs) file.
+
 ## [0.0.2] - 2019-12-21
 
 - Deploy with GitHub Action